Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic[email]==2.9.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
httpx==0.25.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.0.1
PyJWT==2.8.0
aiofiles==23.2.1
firebase-admin==6.2.0
